requirements.txt
numpy==1.14.0
pillow==5.0.0
scipy==1.0.0
